Move TCS/TOU unit tests into test_evaluators_behavior so they run on every PR
Merges the new unit tests (Phase 2 restricted-tool enablement: UNSUPPORTED_TOOLS list, validator accept/reject of newly enabled tools, _stringify_tool_result behavior, end-to-end formatting, real-world SharePoint trace smoke) into the existing _behavior.py files. Per-asset asset.yaml tests_dir already points at those files, so CI picks the new tests up without further changes. The orphan copies under builtin/tests/unit/ (which were not wired into CI) are removed.
